python_playwright_api_framework/requirements.txt
pytest
playwright
pytest-playwright
pytest-xdist
pytest-html
allure-pytest
python-dotenv